remove react-navigation
